export authorization config to utils
