+++ packages/external-db-config/test/drivers/azure_mysql_config_test_support.js
@@ -1,6 +1,7 @@
 const { AzureConfigReader } = require('../../lib/readers/azure_config_reader')
 const Chance = require('chance')
 const chance = new Chance()
+const { validAuthorizationConfig } = require ('../test_utils')
 
 const defineValidConfig = (config) => {
     if (config.host) {
@@ -44,16 +45,6 @@ const validConfigWithAuthorization = () => ({
     ...validConfig(),
     authorization: validAuthorizationConfig.collectionLevelConfig 
 })
-
-const validAuthorizationConfig = {
-    collectionLevelConfig: [
-        {
-            id: chance.word(),
-            readPolicies: ['Admin'],
-            writePolicies: ['Admin'],
-        }
-    ]
-}
 
 const validConfigWithAuthConfig = () => ({
     ...validConfig(),
