+++ packages/external-db-config/test/drivers/gcp_mysql_config_test_support.js
@@ -1,6 +1,7 @@
 const { GcpConfigReader } = require('../../lib/readers/gcp_config_reader')
 const Chance = require('chance')
 const chance = new Chance()
+const { validAuthorizationConfig } = require ('../test_utils')
 
 const defineValidConfig = (config) => {
     if (config.cloudSqlConnectionName) {
@@ -44,16 +45,6 @@ const validConfigWithAuthorization = () => ({
     ...validConfig(),
     authorization: validAuthorizationConfig.collectionLevelConfig 
 })
-
-const validAuthorizationConfig = {
-    collectionLevelConfig: [
-        {
-            id: chance.word(),
-            readPolicies: ['Admin'],
-            writePolicies: ['Admin'],
-        }
-    ]
-}
 
 const validConfigWithAuthConfig = () => ({
     ...validConfig(),
