+++ packages/external-db-config/test/drivers/gcp_spanner_config_test_support.js
@@ -1,6 +1,7 @@
 const { GcpSpannerConfigReader } = require('../../lib/readers/gcp_config_reader')
 const Chance = require('chance')
 const chance = new Chance()
+const { validAuthorizationConfig } = require ('../test_utils')
 
 const defineValidConfig = (config) => {
     if (config.projectId) {
@@ -41,15 +42,6 @@ const validConfigWithAuthorization = () => ({
     authorization: validAuthorizationConfig.collectionLevelConfig 
 })
 
-const validAuthorizationConfig = {
-    collectionLevelConfig: [
-        {
-            id: chance.word(),
-            readPolicies: ['Admin'],
-            writePolicies: ['Admin'],
-        }
-    ]
-}
 const validConfigWithAuthConfig = () => ({
     ...validConfig(),
     auth: {
