+++ packages/external-db-config/test/test_utils.js
@@ -1,0 +1,15 @@
+
+const Chance = require('chance')
+const chance = new Chance()
+
+const validAuthorizationConfig = {
+    collectionLevelConfig: [
+        {
+            id: chance.word(),
+            readPolicies: ['Admin'],
+            writePolicies: ['Admin'],
+        }
+    ]
+}
+
+module.exports = { validAuthorizationConfig }
